cleanup: move semi-external libs to subdir for clarity
