+++ src/coding/circus_decoder.c
@@ -1,5 +1,5 @@
 #include "coding.h"
-#include "circus_decoder_lib.h"
+#include "libs/circus_vq_lib.h"
