+++ src/coding/hca_decoder.c
@@ -1,5 +1,5 @@
 #include "coding.h"
-#include "hca_decoder_clhca.h"
+#include "libs/clhca.h"
 
 
 struct hca_codec_data {
