+++ src/libvgmstream.vcxproj
@@ -92,28 +92,28 @@
     <ClInclude Include="base\mixing_priv.h" />
     <ClInclude Include="base\plugins.h" />
     <ClInclude Include="base\render.h" />
-    <ClInclude Include="coding\acm_decoder_libacm.h" />
-    <ClInclude Include="coding\circus_decoder_lib.h" />
-    <ClInclude Include="coding\circus_decoder_lib_data.h" />
-    <ClInclude Include="coding\circus_decoder_lzxpcm.h" />
     <ClInclude Include="coding\coding.h" />
     <ClInclude Include="coding\coding_utils_samples.h" />
-    <ClInclude Include="coding\compresswave_decoder_lib.h" />
-    <ClInclude Include="coding\g7221_decoder_aes.h" />
-    <ClInclude Include="coding\g7221_decoder_lib.h" />
-    <ClInclude Include="coding\g7221_decoder_lib_data.h" />
     <ClInclude Include="coding\g72x_state.h" />
-    <ClInclude Include="coding\hca_decoder_clhca.h" />
-    <ClInclude Include="coding\ice_decoder_icelib.h" />
     <ClInclude Include="coding\mpeg_decoder.h" />
-    <ClInclude Include="coding\nwa_decoder.h" />
-    <ClInclude Include="coding\relic_decoder_lib.h" />
-    <ClInclude Include="coding\tac_decoder_lib.h" />
-    <ClInclude Include="coding\tac_decoder_lib_data.h" />
-    <ClInclude Include="coding\tac_decoder_lib_ops.h" />
     <ClInclude Include="coding\vorbis_custom_data_fsb.h" />
     <ClInclude Include="coding\vorbis_custom_data_wwise.h" />
     <ClInclude Include="coding\vorbis_custom_decoder.h" />
+    <ClInclude Include="coding\libs\circus_vq_data.h" />
+    <ClInclude Include="coding\libs\circus_vq_lib.h" />
+    <ClInclude Include="coding\libs\circus_vq_lzxpcm.h" />
+    <ClInclude Include="coding\libs\clhca.h" />
+    <ClInclude Include="coding\libs\compresswave_lib.h" />
+    <ClInclude Include="coding\libs\g7221_aes.h" />
+    <ClInclude Include="coding\libs\g7221_data.h" />
+    <ClInclude Include="coding\libs\g7221_lib.h" />
+    <ClInclude Include="coding\libs\icelib.h" />
+    <ClInclude Include="coding\libs\libacm.h" />
+    <ClInclude Include="coding\libs\nwa_lib.h" />
+    <ClInclude Include="coding\libs\relic_lib.h" />
+    <ClInclude Include="coding\libs\tac_data.h" />
+    <ClInclude Include="coding\libs\tac_lib.h" />
+    <ClInclude Include="coding\libs\tac_ops.h" />
     <ClInclude Include="coding\libs\utkdec.h" />
     <ClInclude Include="layout\layout.h" />
     <ClInclude Include="meta\9tav_streamfile.h" />
@@ -212,17 +212,13 @@
     <ClCompile Include="base\render.c" />
     <ClCompile Include="base\seek.c" />
     <ClCompile Include="coding\acm_decoder.c" />
-    <ClCompile Include="coding\acm_decoder_decode.c" />
-    <ClCompile Include="coding\acm_decoder_util.c" />
     <ClCompile Include="coding\adx_decoder.c" />
     <ClCompile Include="coding\asf_decoder.c" />
     <ClCompile Include="coding\atrac9_decoder.c" />
     <ClCompile Include="coding\celt_fsb_decoder.c" />
     <ClCompile Include="coding\circus_decoder.c" />
-    <ClCompile Include="coding\circus_decoder_lib.c" />
     <ClCompile Include="coding\coding_utils.c" />
     <ClCompile Include="coding\compresswave_decoder.c" />
-    <ClCompile Include="coding\compresswave_decoder_lib.c" />
     <ClCompile Include="coding\derf_decoder.c" />
     <ClCompile Include="coding\dpcm_kcej_decoder.c" />
     <ClCompile Include="coding\dsa_decoder.c" />
@@ -237,12 +233,8 @@
     <ClCompile Include="coding\g719_decoder.c" />
     <ClCompile Include="coding\g721_decoder.c" />
     <ClCompile Include="coding\g7221_decoder.c" />
-    <ClCompile Include="coding\g7221_decoder_aes.c" />
-    <ClCompile Include="coding\g7221_decoder_lib.c" />
     <ClCompile Include="coding\hca_decoder.c" />
-    <ClCompile Include="coding\hca_decoder_clhca.c" />
     <ClCompile Include="coding\ice_decoder.c" />
-    <ClCompile Include="coding\ice_decoder_icelib.c" />
     <ClCompile Include="coding\ima_decoder.c" />
     <ClCompile Include="coding\imuse_decoder.c" />
     <ClCompile Include="coding\l5_555_decoder.c" />
@@ -269,13 +261,10 @@
     <ClCompile Include="coding\psx_decoder.c" />
     <ClCompile Include="coding\ptadpcm_decoder.c" />
     <ClCompile Include="coding\relic_decoder.c" />
-    <ClCompile Include="coding\relic_decoder_lib.c" />
-    <ClCompile Include="coding\relic_decoder_mixfft.c" />
     <ClCompile Include="coding\sassc_decoder.c" />
     <ClCompile Include="coding\sdx2_decoder.c" />
     <ClCompile Include="coding\speex_decoder.c" />
     <ClCompile Include="coding\tac_decoder.c" />
-    <ClCompile Include="coding\tac_decoder_lib.c" />
     <ClCompile Include="coding\tantalus_decoder.c" />
     <ClCompile Include="coding\tgcadpcm_decoder.c" />
     <ClCompile Include="coding\ubi_adpcm_decoder.c" />
@@ -293,6 +282,18 @@
     <ClCompile Include="coding\xa_decoder.c" />
     <ClCompile Include="coding\xmd_decoder.c" />
     <ClCompile Include="coding\yamaha_decoder.c" />
+    <ClCompile Include="coding\libs\circus_vq_lib.c" />
+    <ClCompile Include="coding\libs\clhca.c" />
+    <ClCompile Include="coding\libs\compresswave_lib.c" />
+    <ClCompile Include="coding\libs\g7221_aes.c" />
+    <ClCompile Include="coding\libs\g7221_lib.c" />
+    <ClCompile Include="coding\libs\icelib.c" />
+    <ClCompile Include="coding\libs\libacm_decode.c" />
+    <ClCompile Include="coding\libs\libacm_util.c" />
+    <ClCompile Include="coding\libs\nwa_lib.c" />
+    <ClCompile Include="coding\libs\relic_lib.c" />
+    <ClCompile Include="coding\libs\relic_mixfft.c" />
+    <ClCompile Include="coding\libs\tac_lib.c" />
     <ClCompile Include="coding\libs\utkdec.c" />
     <ClCompile Include="layout\blocked.c" />
     <ClCompile Include="layout\blocked_adm.c" />
